+++ lang/src/org/partiql/lang/CompilerPipeline.kt
@@ -17,8 +17,6 @@
 package org.partiql.lang
 
 import com.amazon.ion.IonSystem
-import org.partiql.lang.ast.ExprNode
-import org.partiql.lang.ast.toAstStatement
 import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.eval.Bindings
 import org.partiql.lang.eval.CompileOptions
@@ -109,10 +107,6 @@ interface CompilerPipeline {
 
     /** Compiles the specified PartiQL query using the configured parser. */
     fun compile(query: String): Expression
-
-    @Deprecated("ExprNode is deprecated. Please use PIG generated AST. ")
-    /** Compiles the specified [ExprNode] instance. */
-    fun compile(query: ExprNode): Expression
 
     /** Compiles the specified [PartiqlAst.Statement] instance. */
     fun compile(query: PartiqlAst.Statement): Expression
@@ -264,8 +258,6 @@ internal class CompilerPipelineImpl(
     )
 
     override fun compile(query: String): Expression = compile(parser.parseAstStatement(query))
-
-    override fun compile(query: ExprNode): Expression = compile(query.toAstStatement())
 
     override fun compile(query: PartiqlAst.Statement): Expression {
         val context = StepContext(valueFactory, compileOptions, functions, procedures)