+++ lang/src/org/partiql/lang/ast/meta.kt
@@ -69,3 +69,5 @@ fun MetaContainer.find(tagName: String): Meta? = this[tagName] as Meta?
 fun MetaContainer.hasMeta(tagName: String) = this.containsKey(tagName)
 
 fun MetaContainer.add(meta: Meta): MetaContainer = this.plus(metaContainerOf(meta))
+
+internal typealias IonElementMetaContainer = com.amazon.ionelement.api.MetaContainer