+++ lang/src/org/partiql/lang/eval/EvaluatingCompiler.kt
@@ -26,14 +26,12 @@ import com.amazon.ionelement.api.ionBool
 import com.amazon.ionelement.api.toIonValue
 import org.partiql.lang.ast.AggregateCallSiteListMeta
 import org.partiql.lang.ast.AggregateRegisterIdMeta
-import org.partiql.lang.ast.ExprNode
 import org.partiql.lang.ast.IonElementMetaContainer
 import org.partiql.lang.ast.IsCountStarMeta
 import org.partiql.lang.ast.SourceLocationMeta
 import org.partiql.lang.ast.UniqueNameMeta
 import org.partiql.lang.ast.find
 import org.partiql.lang.ast.sourceLocation
-import org.partiql.lang.ast.toAstStatement
 import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.domains.PartiqlPhysical
 import org.partiql.lang.domains.staticType
@@ -300,9 +298,6 @@ internal class EvaluatingCompiler(
                 }
             )
         }
-
-    @Deprecated("ExprNode is deprecated. Please use PIG generated AST. ")
-    fun compile(exprNode: ExprNode): Expression = compile(exprNode.toAstStatement())
 
     /**
      * Compiles a [PartiqlAst.Statement] tree to an [Expression].