+++ lang/src/org/partiql/lang/syntax/Parser.kt
@@ -17,7 +17,6 @@
 package org.partiql.lang.syntax
 
 import com.amazon.ion.IonSexp
-import org.partiql.lang.ast.ExprNode
 import org.partiql.lang.domains.PartiqlAst
 
 /**
@@ -27,7 +26,4 @@ import org.partiql.lang.domains.PartiqlAst
  */
 interface Parser {
     fun parseAstStatement(source: String): PartiqlAst.Statement
-
-    @Deprecated("Please use parseAstStatement() instead--ExprNode is deprecated.")
-    fun parseExprNode(source: String): ExprNode
 }