+++ lang/test/org/partiql/lang/thread/ThreadInterruptedTests.kt
@@ -12,15 +12,7 @@ import org.junit.jupiter.api.parallel.ExecutionMode
 import org.partiql.lang.CompilerPipeline
 import org.partiql.lang.CompilerPipelineImpl
 import org.partiql.lang.StepContext
-import org.partiql.lang.ast.CaseSensitivity
-import org.partiql.lang.ast.NAry
-import org.partiql.lang.ast.NAryOp
-import org.partiql.lang.ast.ScopeQualifier
-import org.partiql.lang.ast.VariableReference
-import org.partiql.lang.ast.toAstExpr
-import org.partiql.lang.ast.toExprNode
 import org.partiql.lang.domains.PartiqlAst
-import org.partiql.lang.domains.metaContainerOf
 import org.partiql.lang.eval.CompileOptions
 import org.partiql.lang.eval.visitors.VisitorTransformBase
 import org.partiql.lang.syntax.SqlParser
@@ -45,8 +37,6 @@ const val WAIT_FOR_THREAD_TERMINATION_MS: Long = 1000
 @Execution(ExecutionMode.SAME_THREAD)
 class ThreadInterruptedTests {
     private val ion = IonSystemBuilder.standard().build()
-    private val reallyBigNAry = makeBigExprNode(20000000)
-    private val bigNAry = makeBigExprNode(10000000)
     private val bigPartiqlAst = makeBigPartiqlAstExpr(10000000)
 
     /**
@@ -56,16 +46,6 @@ class ThreadInterruptedTests {
      */
     class FakeList<T>(override val size: Int, private val item: T) : AbstractList<T>() {
         override fun get(index: Int): T = item
-    }
-
-    private fun makeBigExprNode(n: Int): NAry {
-        val emptyMetas = metaContainerOf()
-        val variableA = VariableReference("a", CaseSensitivity.INSENSITIVE, ScopeQualifier.UNQUALIFIED, emptyMetas)
-        return NAry(
-            NAryOp.ADD,
-            FakeList(n, variableA),
-            emptyMetas
-        )
     }
 
     private fun makeBigPartiqlAstExpr(n: Int): PartiqlAst.Expr =
@@ -98,28 +78,6 @@ class ThreadInterruptedTests {
             sqlParser.run {
                 endlessTokenList.parseExpression()
             }
-        }
-    }
-
-    @Test
-    fun astChildIterator() {
-        testThreadInterrupt {
-            @Suppress("DEPRECATION")
-            reallyBigNAry.iterator()
-        }
-    }
-
-    @Test
-    fun partiqlAstToExprNode() {
-        testThreadInterrupt {
-            reallyBigNAry.toAstExpr()
-        }
-    }
-
-    @Test
-    fun astToPartiqlAst() {
-        testThreadInterrupt {
-            bigPartiqlAst.toExprNode(ion)
         }
     }